Use a large CTR block during transformations.
