+++ Confuzzle/Confuzzle.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.24720.0
+VisualStudioVersion = 14.0.25123.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Confuzzle", "Confuzzle\Confuzzle.csproj", "{2E571330-9E00-4F06-99EB-68804ED712E5}"
 EndProject
