+++ Confuzzle/Confuzzle/Confuzzle.csproj
@@ -49,12 +49,12 @@
   <ItemGroup>
     <Compile Include="CipherStream.cs" />
     <Compile Include="CipherFactory.cs" />
-    <Compile Include="CtrBlock.cs" />
+    <Compile Include="CtrModeTransform.cs" />
+    <Compile Include="ExtensionMethods.cs" />
     <Compile Include="KeyStretcher.cs" />
     <Compile Include="Options.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="StreamExtensions.cs" />
     <Compile Include="UserAbortException.cs" />
   </ItemGroup>
   <ItemGroup>
